Moved table sort classes out to a separate source file where it belonged to in the first place
